[#86]; refactor: stage3 파일 분리
