+++ src/config/stages/stage3/stage3Bubbles.js
@@ -1,0 +1,6 @@
+/** Stage3 DOM 말풍선(고민 ENTER 힌트 등) */
+
+/** ENTER 말풍선 표시 거리 — 상호작용보다 넓게 */
+export const WORRY_ENTER_HINT_DIST = 7.5;
+export const STAGE3_USER_ENTER_BUBBLE_SHOW_SEC = 3.6;
+export const STAGE3_USER_ENTER_BUBBLE_GAP_SEC = 4.0;
