+++ src/config/stages/stage3/stage3HammerHint.js
@@ -1,0 +1,10 @@
+/**
+ * 글자 위 hammer_only + "click" 말풍선 (현재 미사용 — 재도입 시 Stage3 update에 연동)
+ * @see public/models/stage3/hammer_only.glb
+ */
+export const STAGE3_HAMMER_MODEL_PATH = "/models/stage3/hammer_only.glb";
+export const STAGE3_HAMMER_FLOAT_HEIGHT = 0.25;
+export const STAGE3_HAMMER_FLOAT_AMPLITUDE = 0.06;
+export const STAGE3_HAMMER_FLOAT_SPEED = 2.2;
+export const STAGE3_HAMMER_BASE_ROT_Y = Math.PI * (18 / 180);
+export const STAGE3_HAMMER_BUBBLE_OFFSET_Y = 0.85;
