+++ src/config/stages/stage3/stage3IceCream.js
@@ -1,0 +1,10 @@
+/** Stage3 아이스크림 카트 스폰(cannon-es) 상수 */
+
+/** true면 GLB 대신 디버그 박스만 스폰 */
+export const STAGE3_ICECREAM_DEBUG_BOX_ONLY = false;
+
+/** 아이스크림이 지면에 처음 닿을 때 재생 (랜덤 1종) */
+export const ICECREAM_LAND_SOUND_PATHS = [
+  "/static/sounds/icecream/icecream_sound.mp3",
+  "/static/sounds/icecream/icecream_sound2.mp3",
+];
