+++ src/config/stages/stage3/stage3Keyboard.js
@@ -1,0 +1,19 @@
+/** Stage3 캐릭터 이동 키 (createKeyboardInput용) */
+export const STAGE3_MOVEMENT_KEY_CODES = [
+  "ArrowUp",
+  "ArrowDown",
+  "ArrowLeft",
+  "ArrowRight",
+  "w",
+  "W",
+  "a",
+  "A",
+  "s",
+  "S",
+  "d",
+  "D",
+  "KeyW",
+  "KeyA",
+  "KeyS",
+  "KeyD",
+];
