+++ src/config/stages/stage3/stage3Monitor.js
@@ -1,0 +1,4 @@
+/** Stage3 gum_server 모니터 폴링·Supabase fallback 상수 */
+
+/** REST에서 busy를 한 번도 못 받았을 때만 Supabase fallback (ms) */
+export const STAGE3_MONITOR_FALLBACK_TIMEOUT_MS = 5000;
