+++ src/config/stages/stage3/stage3Overlay.js
@@ -1,0 +1,4 @@
+/** Stage3 React 오버레이(게시판·게임기·텐트) 연동 */
+
+/** App.jsx NoticeModalBoard onClose — 게시판 UI가 닫힌 뒤(Stage3 이스터에그 자막용) */
+export const NOTICE_MODAL_USER_CLOSED_EVENT = "gum:noticeModalClosed";
