+++ src/config/stages/stage3/stage3Portal.js
@@ -1,0 +1,7 @@
+/** Stage3 포탈 → 다음 스테이지 화이트아웃 전환 */
+
+/** 페이드 인 후 풀 화이트 유지 시간(ms), 그다음 stage:switch */
+export const PORTAL_WHITEOUT_FADE_SEC = 1.85;
+export const PORTAL_WHITEOUT_HOLD_MS = 500;
+/** 스테이지 전환 후 새 씬이 드러나도록 화이트 페이드 아웃 */
+export const PORTAL_WHITEOUT_FADE_OUT_SEC = 1.45;
